+++ zcash_history/src/lib.rs
@@ -1,9 +1,13 @@
 //! To be used in Zebra. Originally also consumed by zcashd via FFI bindings,
 //! which reached End-of-Support in July 2026.
 
+#![no_std]
 // Catch documentation errors caused by code changes.
 #![deny(rustdoc::broken_intra_doc_links)]
 #![warn(missing_docs)]
+
+#[macro_use]
+extern crate alloc;
 
 mod entry;
 mod node_data;
@@ -27,8 +31,8 @@ pub enum Error {
     ExpectedNode(Option<EntryLink>),
 }
 
-impl std::fmt::Display for Error {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+impl core::fmt::Display for Error {
+    fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
         match *self {
             Self::ExpectedInMemory(l) => write!(f, "Node/leaf expected to be in memory: {l}"),
             Self::ExpectedNode(None) => write!(f, "Node expected"),
@@ -46,8 +50,8 @@ pub enum EntryLink {
     Generated(u32),
 }
 
-impl std::fmt::Display for EntryLink {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+impl core::fmt::Display for EntryLink {
+    fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
         match *self {
             Self::Stored(v) => write!(f, "stored({v})"),
             Self::Generated(v) => write!(f, "generated({v})"),